refactor: Moved NavBar show status to NavContext
