+++ components/context-api/NavContext.tsx
@@ -3,6 +3,8 @@ import PageWrapper from "../utilities/PageWrapper";
 
 interface NavContextState {
   updateEndpoint: (newEndpoint: string, newPageTitle: string) => void;
+  toggleNavBar: () => void;
+  showNavBar: boolean;
 }
 
 const NavContext = createContext<NavContextState | undefined>(undefined);
@@ -20,16 +22,32 @@ export function NavContextProvider({
 }: React.PropsWithChildren<Record<string, any>>) {
   const [endpoint, setEndpoint] = useState<string>("");
   const [pageTitle, setPageTitle] = useState<string>("");
+  const [showNav, setShowNav] = useState<boolean>(true);
 
   const updateEndpoint = (newEndpoint: string, newPageTitle: string) => {
     setEndpoint(newEndpoint);
     setPageTitle(newPageTitle);
   };
 
+  const toggleNavBar = () => {
+    const navBar = document.querySelector("#navbar");
+    if (!navBar) return;
+
+    const currNavStatus = showNav;
+    setShowNav((prev) => !prev);
+    if (currNavStatus) {
+      navBar.classList.add("hidden");
+    } else {
+      navBar.classList.remove("hidden");
+    }
+  };
+
   return (
     <NavContext.Provider
       value={{
         updateEndpoint,
+        showNavBar: showNav,
+        toggleNavBar,
       }}
     >
       <PageWrapper currPage={endpoint} pageTitle={pageTitle}>
